+++ playservices/src/main/java/no/finntech/android/rx/GoogleApiConnectionSuspended.java
@@ -1,0 +1,9 @@
+package no.finntech.android.rx;
+
+import java.io.IOException;
+
+final class GoogleApiConnectionSuspended extends IOException {
+    public GoogleApiConnectionSuspended(int reason) {
+        super("Reason : " + reason);
+    }
+}
